#8 (**) easier finding of functions etc.. qol changes
